+++ packages/bridge-controller/src/utils/fetch.test.ts
@@ -159,7 +159,7 @@ describe('fetch', () => {
 
       const result = await fetchBridgeQuotes(
         {
-          walletAddress: '0x388c818ca8b9251b393131c08a736a67ccb19297',
+          walletAddress: '0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984',
           srcChainId: 1,
           destChainId: 10,
           srcTokenAddress: AddressZero,
@@ -167,7 +167,7 @@ describe('fetch', () => {
           srcTokenAmount: '20000',
           slippage: 0.5,
           gasIncluded: false,
-          gasless7702: false,
+          gasIncluded7702: false,
         },
         signal,
         BridgeClientId.EXTENSION,
@@ -176,7 +176,7 @@ describe('fetch', () => {
       );
 
       expect(mockFetchFn).toHaveBeenCalledWith(
-        'https://bridge.api.cx.metamask.io/getQuote?walletAddress=0x388C818CA8B9251b393131C08a736A67ccB19297&destWalletAddress=0x388C818CA8B9251b393131C08a736A67ccB19297&srcChainId=1&destChainId=10&srcTokenAddress=0x0000000000000000000000000000000000000000&destTokenAddress=0x0000000000000000000000000000000000000000&srcTokenAmount=20000&insufficientBal=false&resetApproval=false&gasIncluded=false&gasless7702=false&slippage=0.5',
+        'https://bridge.api.cx.metamask.io/getQuote?walletAddress=0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984&destWalletAddress=0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984&srcChainId=1&destChainId=10&srcTokenAddress=0x0000000000000000000000000000000000000000&destTokenAddress=0x0000000000000000000000000000000000000000&srcTokenAmount=20000&insufficientBal=false&resetApproval=false&gasIncluded=false&gasIncluded7702=false&slippage=0.5',
         {
           cacheOptions: {
             cacheRefreshTime: 0,
@@ -213,7 +213,7 @@ describe('fetch', () => {
 
       const result = await fetchBridgeQuotes(
         {
-          walletAddress: '0x388c818ca8b9251b393131c08a736a67ccb19297',
+          walletAddress: '0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984',
           srcChainId: 1,
           destChainId: 10,
           srcTokenAddress: AddressZero,
@@ -221,7 +221,7 @@ describe('fetch', () => {
           srcTokenAmount: '20000',
           slippage: 0.5,
           gasIncluded: false,
-          gasless7702: false,
+          gasIncluded7702: false,
         },
         signal,
         BridgeClientId.EXTENSION,
@@ -230,7 +230,7 @@ describe('fetch', () => {
       );
 
       expect(mockFetchFn).toHaveBeenCalledWith(
-        'https://bridge.api.cx.metamask.io/getQuote?walletAddress=0x388C818CA8B9251b393131C08a736A67ccB19297&destWalletAddress=0x388C818CA8B9251b393131C08a736A67ccB19297&srcChainId=1&destChainId=10&srcTokenAddress=0x0000000000000000000000000000000000000000&destTokenAddress=0x0000000000000000000000000000000000000000&srcTokenAmount=20000&insufficientBal=false&resetApproval=false&gasIncluded=false&gasless7702=false&slippage=0.5',
+        'https://bridge.api.cx.metamask.io/getQuote?walletAddress=0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984&destWalletAddress=0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984&srcChainId=1&destChainId=10&srcTokenAddress=0x0000000000000000000000000000000000000000&destTokenAddress=0x0000000000000000000000000000000000000000&srcTokenAmount=20000&insufficientBal=false&resetApproval=false&gasIncluded=false&gasIncluded7702=false&slippage=0.5',
         {
           cacheOptions: {
             cacheRefreshTime: 0,
@@ -285,7 +285,7 @@ describe('fetch', () => {
 
       const result = await fetchBridgeQuotes(
         {
-          walletAddress: '0x388c818ca8b9251b393131c08a736a67ccb19297',
+          walletAddress: '0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984',
           srcChainId: 1,
           destChainId: 10,
           srcTokenAddress: AddressZero,
@@ -293,7 +293,7 @@ describe('fetch', () => {
           srcTokenAmount: '20000',
           slippage: 0.5,
           gasIncluded: false,
-          gasless7702: false,
+          gasIncluded7702: false,
         },
         signal,
         BridgeClientId.EXTENSION,
@@ -302,7 +302,7 @@ describe('fetch', () => {
       );
 
       expect(mockFetchFn).toHaveBeenCalledWith(
-        'https://bridge.api.cx.metamask.io/getQuote?walletAddress=0x388C818CA8B9251b393131C08a736A67ccB19297&destWalletAddress=0x388C818CA8B9251b393131C08a736A67ccB19297&srcChainId=1&destChainId=10&srcTokenAddress=0x0000000000000000000000000000000000000000&destTokenAddress=0x0000000000000000000000000000000000000000&srcTokenAmount=20000&insufficientBal=false&resetApproval=false&gasIncluded=false&gasless7702=false&slippage=0.5',
+        'https://bridge.api.cx.metamask.io/getQuote?walletAddress=0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984&destWalletAddress=0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984&srcChainId=1&destChainId=10&srcTokenAddress=0x0000000000000000000000000000000000000000&destTokenAddress=0x0000000000000000000000000000000000000000&srcTokenAmount=20000&insufficientBal=false&resetApproval=false&gasIncluded=false&gasIncluded7702=false&slippage=0.5',
         {
           cacheOptions: {
             cacheRefreshTime: 0,
@@ -349,7 +349,7 @@ describe('fetch', () => {
 
       const result = await fetchBridgeQuotes(
         {
-          walletAddress: '0x388c818ca8b9251b393131c08a736a67ccb19297',
+          walletAddress: '0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984',
           srcChainId: 1,
           destChainId: 10,
           srcTokenAddress: AddressZero,
@@ -357,7 +357,7 @@ describe('fetch', () => {
           srcTokenAmount: '20000',
           slippage: 0.5,
           gasIncluded: false,
-          gasless7702: false,
+          gasIncluded7702: false,
           aggIds: ['socket', 'lifi'],
           bridgeIds: ['bridge1', 'bridge2'],
           noFee: true,
@@ -369,7 +369,7 @@ describe('fetch', () => {
       );
 
       expect(mockFetchFn).toHaveBeenCalledWith(
-        'https://bridge.api.cx.metamask.io/getQuote?walletAddress=0x388C818CA8B9251b393131C08a736A67ccB19297&destWalletAddress=0x388C818CA8B9251b393131C08a736A67ccB19297&srcChainId=1&destChainId=10&srcTokenAddress=0x0000000000000000000000000000000000000000&destTokenAddress=0x0000000000000000000000000000000000000000&srcTokenAmount=20000&insufficientBal=false&resetApproval=false&gasIncluded=false&gasless7702=false&slippage=0.5&noFee=true&aggIds=socket%2Clifi&bridgeIds=bridge1%2Cbridge2',
+        'https://bridge.api.cx.metamask.io/getQuote?walletAddress=0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984&destWalletAddress=0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984&srcChainId=1&destChainId=10&srcTokenAddress=0x0000000000000000000000000000000000000000&destTokenAddress=0x0000000000000000000000000000000000000000&srcTokenAmount=20000&insufficientBal=false&resetApproval=false&gasIncluded=false&gasIncluded7702=false&slippage=0.5&noFee=true&aggIds=socket%2Clifi&bridgeIds=bridge1%2Cbridge2',
         {
           cacheOptions: {
             cacheRefreshTime: 0,