+++ packages/gator-permissions-controller/src/index.ts
@@ -7,6 +7,7 @@ export type {
   GatorPermissionsControllerState,
   GatorPermissionsControllerMessenger,
   GatorPermissionsControllerGetStateAction,
+  GatorPermissionsControllerDecodePermissionFromPermissionContextForOriginAction,
   GatorPermissionsControllerFetchAndUpdateGatorPermissionsAction,
   GatorPermissionsControllerEnableGatorPermissionsAction,
   GatorPermissionsControllerDisableGatorPermissionsAction,
@@ -14,19 +15,12 @@ export type {
   GatorPermissionsControllerEvents,
   GatorPermissionsControllerStateChangeEvent,
 } from './GatorPermissionsController';
+export type { DecodedPermission } from './decodePermission';
 export type {
   GatorPermissionsControllerErrorCode,
   GatorPermissionsSnapRpcMethod,
-  MetaMaskBasePermissionData,
-  NativeTokenStreamPermission,
-  NativeTokenPeriodicPermission,
-  Erc20TokenStreamPermission,
-  Erc20TokenPeriodicPermission,
   CustomPermission,
-  PermissionTypes,
-  AccountSigner,
-  WalletSigner,
-  SignerParam,
+  PermissionTypesWithCustom,
   PermissionRequest,
   PermissionResponse,
   PermissionResponseSanitized,
@@ -36,4 +30,16 @@ export type {
   SupportedGatorPermissionType,
   GatorPermissionsMapByPermissionType,
   GatorPermissionsListByPermissionTypeAndChainId,
+  DelegationDetails,
 } from './types';
+
+export type {
+  NativeTokenStreamPermission,
+  NativeTokenPeriodicPermission,
+  Erc20TokenStreamPermission,
+  Erc20TokenPeriodicPermission,
+  AccountSigner,
+  WalletSigner,
+  Signer,
+  MetaMaskBasePermissionData,
+} from '@metamask/7715-permission-types';